fix(transform-eventual-send): split out rewriter.js to fix cycle
The old default export (used for SES1's evaluator transforms array) is now in
a separate file:

```js
import makeEventualSendTransformer from '@agoric/transform-eventual-send/src/rewriter';
```

The new smaller `makeTransform` function (used by bundle-source) is a named
export of the package:

```js
import { makeTransform } from '@agoric/transform-eventual-send';
```

This breaks the dependency cycle between bundle-source (which needs
`makeTransform`) and `makeEventualSendTransformer` (which needs a file
created at `yarn build` time by running bundle-source).
