+++ packages/default-evaluate-options/src/index.js
@@ -1,6 +1,6 @@
 import { parse } from '@agoric/babel-parser';
 import generate from '@babel/generator';
-import makeEventualSendTransformer from '@agoric/transform-eventual-send';
+import makeEventualSendTransformer from '@agoric/transform-eventual-send/src/rewriter';
 
 export default function makeDefaultEvaluateOptions() {
   const shims = [];
